Fourth attempt

+++ scripts/trainer.py
@@ -2118,7 +2118,7 @@ def main():
                                 are_we_v_pred = True
                             elif args.scale_v_pred_loss:
                                 are_we_v_pred = True
-                            starting_loss = (target - model_pred) ** 2
+                            starting_loss = (target.float() - model_pred.float()) ** 2
                             loss = tu.apply_snr_weight_neo(are_we_v_pred, starting_loss, timesteps, noise_scheduler, args.min_snr_gamma, accelerator)
                             loss = loss.mean()
                         else:
